refactor: move loading-related components to separate folder
